+++ lib/flutter_nested_categories.dart
@@ -1,0 +1,10 @@
+/// Flutter Nested Categories
+///
+/// A Flutter package that allows you to create nested categories that
+/// are very easy to use and customize.
+library flutter_nested_categories;
+
+export "src/category.dart";
+export "src/category_header_capitalization.dart";
+export "src/category_header_styling.dart";
+export "src/category_list.dart";